requirements.txt
fastapi==0.120.3
uvicorn==0.38.0
pandas
numpy
requests
typing
pydantic
google-cloud-aiplatform
google-genai
bs4
pyspellchecker
fuzzywuzzy
levenshtein
aiohttp